requirements.txt
# 基本ライブラリ（Phase 1-3）
opencv-python>=4.8.0      # 画像処理・コンピュータビジョン
numpy>=1.24.0             # 数値計算・配列操作
scipy>=1.10.0             # 科学計算・統計処理
matplotlib>=3.7.0         # 基本的なグラフ生成
pillow>=10.0.0            # 画像I/O・基本処理
pandas>=2.0.0             # データ操作・統計
seaborn>=0.12.0           # 統計的可視化
scikit-image>=0.20.0      # 画像解析・特徴抽出

# 高度な色彩科学（Phase 4以降）
colour-science>=0.4.0     # 色彩科学計算・色差測定
colorspacious>=1.1.0      # 色空間変換

# 高度な画像解析（Phase 4以降）
pywavelets>=1.4.0         # ウェーブレット変換（周波数解析）
mahotas>=1.4.0            # コンピュータビジョン

# 機械学習（印象分析用）
scikit-learn>=1.3.0       # 機械学習アルゴリズム

# その他
plotly>=5.15.0            # インタラクティブ可視化
pyyaml>=6.0               # YAML設定ファイル処理